Bugfix to nimble plot

+++ singlecell/src/org/labkey/singlecell/run/NimbleHelper.java
@@ -505,30 +505,37 @@ public class NimbleHelper
 
             resultMap.put(genome, reportResultsGz);
 
-            // Also run nimble plot. Always re-run since this is fast:
-            List<String> plotArgs = new ArrayList<>();
-            plotArgs.add("python3");
-            plotArgs.add("-m");
-            plotArgs.add("nimble");
-
-            plotArgs.add("plot");
-            plotArgs.add("--input_file");
-            plotArgs.add("/work/" + reportResultsGz.getName());
-
-            File plotResultsHtml = getReportHtmlFileFromResults(reportResultsGz);
-            if (reportResultsGz.exists())
+            if (SequencePipelineService.get().hasMinLineCount(reportResultsGz, 2))
             {
-                plotResultsHtml.delete();
+                // Also run nimble plot. Always re-run since this is fast:
+                List<String> plotArgs = new ArrayList<>();
+                plotArgs.add("python3");
+                plotArgs.add("-m");
+                plotArgs.add("nimble");
+
+                plotArgs.add("plot");
+                plotArgs.add("--input_file");
+                plotArgs.add("/work/" + reportResultsGz.getName());
+
+                File plotResultsHtml = getReportHtmlFileFromResults(reportResultsGz);
+                if (reportResultsGz.exists())
+                {
+                    plotResultsHtml.delete();
+                }
+
+                plotArgs.add("--output_file");
+                plotArgs.add("/work/" + plotResultsHtml.getName());
+
+                runUsingDocker(plotArgs, output, null);
+
+                if (!plotResultsHtml.exists())
+                {
+                    throw new PipelineJobException("Missing file: " + plotResultsHtml.getPath());
+                }
             }
-
-            plotArgs.add("--output_file");
-            plotArgs.add("/work/" + plotResultsHtml.getName());
-
-            runUsingDocker(plotArgs, output, null);
-
-            if (!plotResultsHtml.exists())
+            else
             {
-                throw new PipelineJobException("Missing file: " + plotResultsHtml.getPath());
+                getPipelineCtx().getLogger().info("Only single line found in results, skipping nimble plot");
             }
         }
